Swift's numbers take plain numbers: `new NSNumber(42)`, `{ at: 0 }`, no casts
The third step of the Swift-shaped surface. A scalar brand may be optional --
`number & { readonly __c_double?: true }` -- and a plain `number` is then
assignable to it while the call still carries the brand's C type. objc:types
exports Swift's names that way: `Double`, `Float`, `CGFloat`, `TimeInterval`,
`Int`, `UInt`, and `Int8` through `UInt64`.

`Int` is 64 bits over a `number`, so an integer past 2^53 rounds, as it does
in any bridge to JavaScript; the required `bigint` brands in c:types are
unchanged for code that needs every bit. A value keeps its brand's
representation inside the program, as every branded number here does, so
`Int` arithmetic is integer arithmetic, which is Swift's.

A labels literal's placeholder is typed at the parameter's declared type:
`{ at: 0 }` is `{ at: number }`, the parameter `{ at: UInt }`, and the
argument is converted to the parameter before any role reads it.

macos-classes has no cast left and matches its oracle on both backends; a
test pins that a send taking `Int`, `CGFloat` and `UInt` from plain numbers
is typed `long`, `double`, `unsigned long`.

+++ examples/interop/macos-classes/src/main.ts
@@ -4,7 +4,7 @@
 // against the same program in Objective-C (`reference/classes.m`).
 import { NSMutableArray, NSNumber, NSObject, NSOperation, NSProcessInfo, NSString } from "objc:Foundation";
 import { report, report_string, weak_alive, weak_watch } from "c:support";
-import type { c_int, c_ulong } from "c:types";
+import type { c_int } from "c:types";
 
 let watch = 0 as c_int;
 
@@ -19,16 +19,16 @@ function main(): void {
   const list = new NSMutableArray();
   report(`empty ${list.count}`);
   for (const n of [1, 2, 3]) {
-    list.addObject(new NSNumber(n as c_int));
+    list.addObject(new NSNumber(n));
   }
   report(`count ${list.count}`);
-  list.insert(new NSNumber(0 as c_int), { at: 0n as c_ulong });
-  report(`inserted ${list.count} first ${(list.object(0n as c_ulong) as NSNumber).intValue}`);
+  list.insert(new NSNumber(0), { at: 0 });
+  report(`inserted ${list.count} first ${(list.object(0) as NSNumber).intValue}`);
 
-  const answer = new NSNumber(42 as c_int);
-  report(`number ${answer.intValue} equal ${answer.isEqual(new NSNumber(42 as c_int))}`);
+  const answer = new NSNumber(42);
+  report(`number ${answer.intValue} equal ${answer.isEqual(new NSNumber(42))}`);
   report(`kinds ${answer instanceof NSNumber} ${answer instanceof NSString} ${list instanceof NSObject}`);
-  report(`processors ${NSProcessInfo.processInfo.processorCount > 0n}`);
+  report(`processors ${NSProcessInfo.processInfo.processorCount > 0}`);
 
   const text = new NSString("worker");
   report(`length ${text.length}`);
